\name{graph_demo}
\alias{graph_demo}
\docType{data}
\title{Demo graph}
\description{An example of undirected graph for execution with stability function.}
\usage{
	data(graph_demo)
}
\format{
	\code{graph_demo} is a weighted adjacency matrix with 640 nodes.
}
\seealso{
	\code{\link{graph_rings}},
	\code{\link{stability}}
}
\keyword{datasets}
