\name{graph_rings}
\alias{graph_rings}
\docType{data}
\title{Ring of rings graph}
\description{An example of undirected graph for execution with stability function.}
\usage{
	data(graph_rings)
}
\format{
	\code{graph_rings} is a weighted adjacency matrix with 100 nodes.
}
\seealso{
	\code{\link{graph_demo}},
	\code{\link{stability}}
}
\keyword{datasets}
