\name{StabilityR-package}
\alias{StabilityR-package}
\alias{StabilityR}
\docType{package}
\title{\packageTitle{StabilityR}}
\description{\packageDescription{StabilityR}}
\details{
The DESCRIPTION file:
\packageDESCRIPTION{StabilityR}
\packageIndices{StabilityR}
}
\author{
\packageAuthor{StabilityR}
Maintainer: \packageMaintainer{StabilityR}
}
% \references{
% }
\keyword{package}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/tleja/StabilityR}
  \item Report bugs at \url{https://github.com/tleja/StabilityR/issues}
}
}
