\name{StabilityR-package}
\alias{StabilityR-package}
\alias{StabilityR}
\docType{package}
\title{
\packageTitle{StabilityR}
}
\description{
\packageDescription{StabilityR}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{StabilityR}
\packageIndices{StabilityR}
%~~ An overview of how to use the package, including the most important ~~
%~~ functions ~~
}
\author{
\packageAuthor{StabilityR}

Maintainer: \packageMaintainer{StabilityR}
}
%\references{
%~~ Literature or other references for background information ~~
%}
%~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
%~~ the R documentation directory ~~
%\keyword{ package }
%\seealso{
%~~ Optional links to other man pages, e.g. ~~
%~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
%}
%\examples{
%~~ simple examples of the most important functions ~~
%}
